+++ CHANGELOG.md
@@ -1,8 +1,20 @@
-# 3.0
+# 4.0.0
+
+- Remove 
+  - prop `static`  
+    No need for static now 
+  - prop `slotTag`, `slotClass`, `slotId`
+
+- Add
+  - expose `shadow_root`
+  - component `ShadowStyle`  
+    `ShadowRoot.Style` or `ShadowStyle`
+
+# 3.0.0
 
 Make version number same to vue  
 ./dist files name same to vue
 
-# 2.0
+# 2.0.0
 
 Support vue 3